requirements.txt
tensorflow>=2.16.0
python-chess>=1.10.0
numpy>=1.24.0
tqdm>=4.65.0
